requirements.txt
fastapi==0.111.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
Pillow==10.3.0
pydantic==2.7.3 
pdf2image==1.17.0
pytesseract==0.3.10